+++ src/core/client/views/inventory/html/app.js
@@ -19,7 +19,19 @@ const app = new Vue({
             pageIndex: 0,
             pageMeta: ['Tab 1', 'Tab 2', 'Tab 3', 'Tab 4', 'Tab 5'],
             pageIcons: ['icon-box', 'icon-box', 'icon-box', 'icon-box', 'icon-box'],
-            equipmentMeta: ['Hat', 'Mask', 'Shirt', 'Pants', 'Feet', 'Glasses', 'Ears', 'Bag', 'Armour'],
+            equipmentMeta: [
+                'Hat',
+                'Mask',
+                'Shirt',
+                'Pants',
+                'Feet',
+                'Glasses',
+                'Ears',
+                'Bag',
+                'Armour',
+                'Watches',
+                'Bracelets'
+            ],
             inventory: [[], [], [], [], [], []],
             ground: [],
             equipment: [],
@@ -66,7 +78,7 @@ const app = new Vue({
             this.ground = newGround;
         },
         updateEquipment(equipmentItems) {
-            const newEquipment = new Array(9).fill(null);
+            const newEquipment = new Array(11).fill(null);
             equipmentItems.forEach((item) => {
                 if (!item) {
                     return;
@@ -441,7 +453,7 @@ const app = new Vue({
         // Used to populate data on entry.
         this.inventory = new Array(6).fill(new Array(28).fill(null));
         this.ground = new Array(8).fill(null);
-        this.equipment = new Array(9).fill(null);
+        this.equipment = new Array(11).fill(null);
         this.toolbar = new Array(4).fill(null);
 
         if ('alt' in window) {