+++ documentation/changelog.md
@@ -19,6 +19,10 @@ description: List of all the changes and current versions.
 + Added Forced Camera Shake on Aim
 + Added Shared Config for Camera Shake on Aim
 + Added Support for Consumables on the Toolbar
++ Added Watches & Bracelets to Inventory
++ Added Clothing Shop Interface
++ Added Clothing Synchronization
++ Added Clothing Item Generator
 ```
 
 ## 1.3.0 - January 23-26, 2021